requirements.txt
alembic==1.16.5
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
greenlet==3.3.2
psycopg[binary]==3.2.10
redis==6.4.0
pydantic-settings==2.10.1
firebase-admin==7.2.0
python-multipart==0.0.20
boto3>=1.34,<2
sqladmin==0.20.1
itsdangerous==2.2.0
httpx==0.28.1
websockets==16.0
feedparser==6.0.11
cryptography==46.0.5
prometheus-client==0.23.1
cloudscraper==1.2.71
